+++ Sources/DefaultsAdapter.swift
@@ -48,6 +48,12 @@ public struct DefaultsAdapter<KeyStore: DefaultsKeyStore> {
 
     #if !os(Linux) && !os(watchOS)
     internal let syncer: DefaultsSyncer
+
+    internal init(defaults: UserDefaults, keyStore: KeyStore, remoteStore: RemoteStore) {
+        self.defaults = defaults
+        self.keyStore = keyStore
+        self.syncer = DefaultsSyncer(defaults: defaults, remoteStore: remoteStore)
+    }
     #endif
 
     public init(defaults: UserDefaults, keyStore: KeyStore) {
@@ -55,7 +61,7 @@ public struct DefaultsAdapter<KeyStore: DefaultsKeyStore> {
         self.keyStore = keyStore
 
         #if !os(Linux) && !os(watchOS)
-        self.syncer = DefaultsSyncer(defaults: defaults)
+        self.syncer = DefaultsSyncer(defaults: defaults, remoteStore: NSUbiquitousKeyValueStore.default)
         #endif
     }